+++ src/products/products.service.ts
@@ -1,14 +1,26 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, forwardRef, Inject, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Product, ProductDocument } from './schemas/producct.schema';
 import { Model } from 'mongoose';
+import { CreateProductDto } from './dto/create-product.dto';
+import { NotFoundError } from 'rxjs';
+import { CategoriesService } from 'src/categories/categories.service';
+import { Order, OrderDocument } from 'src/orders/schemas/order.schema';
 
 @Injectable()
 export class ProductsService {
-    constructor(@InjectModel(Product.name) private productModel: Model<ProductDocument>) {}
+    constructor(@InjectModel(Product.name) private productModel: Model<ProductDocument>,
+    @InjectModel(Order.name) private orderModel: Model<OrderDocument>,
+    @Inject(forwardRef(() => CategoriesService))
+    private categoriesService: CategoriesService,
+    ) {}
 
-    async create(productData: Partial<Product>) {
-    const newProduct = new this.productModel(productData);
+  async create(dto: CreateProductDto) {
+    const category = await this.categoriesService.findOne(dto.categoryId);
+    if (!category) {
+      throw new NotFoundError('Category not found');
+    }
+    const newProduct = new this.productModel(dto);
     return newProduct.save();
   }
 
@@ -29,7 +41,20 @@ export class ProductsService {
   }
 
   async remove(id: string) {
+    const order = await this.orderModel.findOne({ 
+      'items.productId': id 
+    });
+
+    if (order){
+      throw new BadRequestException(
+        'Cannot delete product used in orders',
+      )
+    }
     return this.productModel.findByIdAndDelete(id);
+  }
+
+  async findByCategory(categoryId: string) {
+    return this.productModel.find({ categoryId });
   }
 
 }
